refactor homepage: delete old version, ajust structure, add guest notification
