+++ client/src/index.css
@@ -118,6 +118,8 @@ body {
 
 .page-content {
   flex: 1;
+  display: flex;
+  flex-direction: column;
 }
 
 header,
